BUILDS: Things are probably broken, as i have changed many timing variables around.
Still need to work some things out, but the time classes/concepts seem like they might work.

+++ src/Animation/AnimationSystem.cpp
@@ -40,7 +40,7 @@ namespace Core
 		DEBUG_SHUTDOWN(AnimationSystem);
 	}
 
-	void AnimationSystem::update(uint64_t deltaTime)
+	void AnimationSystem::update(uint32_t deltaTime)
 	{
 		m_timer.updateBy(deltaTime);
